requirements.txt
berserk==0.10.0
Deprecated==1.2.12
et-xmlfile==1.0.1
ndjson==0.3.1
openpyxl==3.0.7
python-lichess==0.9
